fix(test): address local review findings
