+++ packages/morpho-sdk/src/entities/vaultV1/vaultV1.test.ts
@@ -1,0 +1,61 @@
+import { createPublicClient, http } from "viem";
+import { mainnet } from "viem/chains";
+import { describe, expect, test } from "vitest";
+import { SteakhouseUsdcVaultV1 } from "../../../test/fixtures/vaultV1.js";
+import { morphoViemExtension } from "../../client/index.js";
+import { ChainIdMismatchError } from "../../types/index.js";
+
+describe("MorphoVaultV1 chain validation", () => {
+  test("getData throws ChainIdMismatchError when client chain differs", async () => {
+    const publicClient = createPublicClient({
+      chain: mainnet,
+      transport: http("https://rpc.example"),
+    });
+    const vault = publicClient
+      .extend(morphoViemExtension())
+      .morpho.vaultV1(SteakhouseUsdcVaultV1.address, mainnet.id + 1);
+
+    await expect(vault.getData()).rejects.toThrow(ChainIdMismatchError);
+  });
+
+  test("deposit throws ChainIdMismatchError when client chain differs", () => {
+    const publicClient = createPublicClient({
+      chain: mainnet,
+      transport: http("https://rpc.example"),
+    });
+    const vault = publicClient
+      .extend(morphoViemExtension())
+      .morpho.vaultV1(SteakhouseUsdcVaultV1.address, mainnet.id + 1);
+
+    expect(() =>
+      vault.deposit({
+        amount: 1n,
+        userAddress: SteakhouseUsdcVaultV1.address,
+        vaultData: {} as never,
+      }),
+    ).toThrow(ChainIdMismatchError);
+  });
+
+  test("withdraw and redeem throw ChainIdMismatchError when client chain differs", () => {
+    const publicClient = createPublicClient({
+      chain: mainnet,
+      transport: http("https://rpc.example"),
+    });
+    const vault = publicClient
+      .extend(morphoViemExtension())
+      .morpho.vaultV1(SteakhouseUsdcVaultV1.address, mainnet.id + 1);
+
+    expect(() =>
+      vault.withdraw({
+        amount: 1n,
+        userAddress: SteakhouseUsdcVaultV1.address,
+      }),
+    ).toThrow(ChainIdMismatchError);
+    expect(() =>
+      vault.redeem({
+        shares: 1n,
+        userAddress: SteakhouseUsdcVaultV1.address,
+      }),
+    ).toThrow(ChainIdMismatchError);
+  });
+});
