+++ packages/morpho-test/package.json
@@ -31,7 +31,7 @@
   "peerDependencies": {
     "@morpho-org/blue-sdk": "^6.0.0",
     "@morpho-org/morpho-ts": "^2.7.0",
-    "@morpho-org/test": "^2.8.0",
+    "@morpho-org/test": "^2.9.0",
     "viem": "^2.21.10"
   },
   "devDependencies": {
